# Optimize source mapping into external source map sources