Fixes #40054

## Timings

| fn                                       | Before  | After | Improvement
| ------------------------- | -------|------|--------|
| emit phase                        | 13.4      | 8.8    | -34% |
| `emitSourcePos`              | 4.8s      | 0.6s  | -87% |
| `setSourceMapSource`    | 4.3s      | 0.1s  | -98% |