@DanielRosenwasser 2020-08-19
Shucks, we really should have run a perf test on this.

@typescript-bot perf test this